Implement giveNumber and giveOperator
These two functions modify the input display such. If a number button was pressed, giveNumber will append the input display with the number whose button was pressed.

If an operator button was pressed, giveOperator will append the input display with the operator whose button was pressed if there is no operator in the input yet, or replace the current operator if there is an operator in the input already.