Simplify giveOperator
The function giveOperator was pretty unwieldy. It had separate logic if there was a result, if there was an input without or without a second number, and if there was only one number. By calling giveEquals if there is an operator and adding the new operator to that result, we handle several of these cases in one conditional.